[Refector] : 금액에 콤마추가 메서드 Util로 이동
