+++ app/src/main/java/kr/sparta/tripmate/util/method/method.kt
@@ -28,3 +28,25 @@ fun removeHtmlTags(input: String): String {
 fun Int.toMoneyFormat(): String = DecimalFormat("#,###").format(this)
 fun Int.toTimeFormat(): String = DecimalFormat("00").format(this)
 
+/**
+ * 작성자: 서정한
+ * 내용: 금액에 콤마 추가
+ * */
+fun setCommaForMoneeyText(money: String): String {
+    val commaCount = if (money.length % 3 == 0) {
+        (money.length / 3) - 1
+    } else {
+        money.length / 3
+    }
+    var commaStartIndex = money.length % 3
+    val comma = ','
+    val stringBuilder = StringBuilder()
+    stringBuilder.append(money)
+
+    for (i in 0 until commaCount) {
+        stringBuilder.insert(commaStartIndex, comma)
+        commaStartIndex += 4
+    }
+
+    return "${stringBuilder}원"
+}
